+++ examples/card_test/index.js
@@ -5,7 +5,7 @@ import { ModelRoot, ViewRoot, StartWorldcore, Actor, Pawn, mix, InputManager, Pl
     AM_Predictive, PM_Predictive,
     AM_PointerTarget, PM_Pointer, PM_PointerTarget, CardActor, CardPawn,
     q_axisAngle, m4_rotationQ, m4_identity, GetPawn, WidgetActor, WidgetPawn, ImageWidgetPawn, CanvasWidgetPawn, ImageWidgetActor, CanvasWidgetActor,
-    TextWidgetActor, ButtonWidgetActor, GetViewService } from "@croquet/worldcore";
+    TextWidgetActor, ButtonWidgetActor, GetViewService, UIManager, ButtonWidget, TextWidget, Widget, AM_Smoothed, PM_Smoothed, PM_Driver, v3_scale, v3_add, TAU, v3_rotate, toDeg, q_multiply, m4_multiply, m4_scaleRotationTranslation } from "@croquet/worldcore";
 
 import { Widget3, VisibleWidget3, ControlWidget3, PM_Widget3, PM_WidgetPointer, WidgetManager, ImageWidget3, CanvasWidget3, TextWidget3, ButtonWidget3, ToggleWidget3, ToggleSet3, SliderWidget3, BoxWidget3, DragWidget3 } from "./ThreeWidget";
 
@@ -18,7 +18,7 @@ import kwark from "./assets/kwark.otf";
 //-- MyAvatar ------------------------------------------------------------------------------
 //------------------------------------------------------------------------------------------
 
-class MyAvatar extends mix(Actor).with(AM_Predictive, AM_Player) {
+class MyAvatar extends mix(Actor).with(AM_Smoothed, AM_Player) {
 
     get pawn() {return AvatarPawn}
     get color() {return this._color || [1,1,1,1]}
@@ -30,12 +30,13 @@ MyAvatar.register('MyAvatar');
 //-- AvatarPawn ----------------------------------------------------------------------------
 //------------------------------------------------------------------------------------------
 
-class AvatarPawn extends mix(Pawn).with(PM_Predictive, PM_Player, PM_ThreeVisible, PM_ThreeCamera, PM_WidgetPointer) {
+class AvatarPawn extends mix(Pawn).with(PM_Smoothed, PM_Driver, PM_Player, PM_ThreeVisible, PM_ThreeCamera, PM_WidgetPointer) {
     constructor(...args) {
         super(...args);
 
-        this.fore = this.back = this.left = this.right = 0;
-        this.ccw = this.cw = 0;
+        this.fore = this.back = this.left = this.right = this.pitch = this.yaw = 0;
+        this.speed = 5;
+        this.turnSpeed = 0.002;
 
         this.geometry = new THREE.BoxGeometry( 1, 1, 1 );
         this.material = new THREE.MeshStandardMaterial( {color: new THREE.Color(...this.actor.color)} );
@@ -43,21 +44,29 @@ class AvatarPawn extends mix(Pawn).with(PM_Predictive, PM_Player, PM_ThreeVisibl
         this.setRenderObject(cube);
 
         if (this.isMyPlayerPawn) {
-            this.subscribe("input", "wDown", () => {this.fore = 1; this.changeVelocity()});
-            this.subscribe("input", "wUp", () => {this.fore = 0; this.changeVelocity()});
-            this.subscribe("input", "sDown", () => {this.back = 1; this.changeVelocity()});
-            this.subscribe("input", "sUp", () => {this.back = 0; this.changeVelocity()});
-            this.subscribe("input", "qDown", () => {this.left = 1; this.changeVelocity()});
-            this.subscribe("input", "qUp", () => {this.left = 0; this.changeVelocity()});
-            this.subscribe("input", "eDown", () => {this.right = 1; this.changeVelocity()});
-            this.subscribe("input", "eUp", () => {this.right = 0; this.changeVelocity()});
+            this.subscribe("input", "wDown", () => {this.fore = 1});
+            this.subscribe("input", "wUp", () => {this.fore = 0});
+            this.subscribe("input", "sDown", () => {this.back = 1});
+            this.subscribe("input", "sUp", () => {this.back = 0});
 
-            this.subscribe("input", "aDown", () => {this.ccw = 1; this.changeSpin()});
-            this.subscribe("input", "aUp", () => {this.ccw = 0; this.changeSpin()});
-            this.subscribe("input", "dDown", () => {this.cw = 1; this.changeSpin()});
-            this.subscribe("input", "dUp", () => {this.cw = 0; this.changeSpin()});
+            this.subscribe("input", "ArrowUpDown", () => {this.fore = 1});
+            this.subscribe("input", "ArrowUpUp", () => {this.fore = 0});
+            this.subscribe("input", "ArrowDownDown", () => {this.back = 1});
+            this.subscribe("input", "ArrowDownUp", () => {this.back = 0});
 
-            // this.subscribe("input", "pointerDown", this.doPointerDown);
+            this.subscribe("input", "aDown", () => {this.left = 1});
+            this.subscribe("input", "aUp", () => {this.left = 0});
+            this.subscribe("input", "dDown", () => {this.right = 1});
+            this.subscribe("input", "dUp", () => {this.right = 0});
+
+            this.subscribe("input", "ArrowLeftDown", () => {this.left = 1});
+            this.subscribe("input", "ArrowLeftUp", () => {this.left = 0});
+            this.subscribe("input", "ArrowRightDown", () => {this.right = 1});
+            this.subscribe("input", "ArrowRightUp", () => {this.right = 0});
+
+            this.subscribe("input", "pointerDown", this.doPointerDown);
+            this.subscribe("input", "pointerUp", this.doPointerUp);
+            this.subscribe("input", "pointerDelta", this.doPointerDelta);
 
         }
 
@@ -69,17 +78,53 @@ class AvatarPawn extends mix(Pawn).with(PM_Predictive, PM_Player, PM_ThreeVisibl
         this.material.dispose();
     }
 
-    changeVelocity() {
-        const velocity = [ -0.01 * (this.left - this.right), 0,  -0.01 * (this.fore - this.back)]
-        this.setVelocity(velocity);
+    get velocity() {
+        return [ (this.left - this.right), 0,  (this.fore - this.back)];
     }
 
-    changeSpin() {
-        const spin = q_axisAngle([0,1,0], 0.001 * (this.ccw - this.cw) )
-        this.setSpin(spin);
+    update(time, delta) {
+        super.update(time,delta);
+        const pitchQ = q_axisAngle([1,0,0], this.pitch);
+        const yawQ = q_axisAngle([0,1,0], this.yaw);
+        // const lookQ = q_multiply(pitchQ, yawQ);
+        const v = v3_scale(this.velocity, -this.speed * delta/1000)
+        const v2 = v3_rotate(v, yawQ);
+        const t = v3_add(this.translation, v2)
+        this.positionTo(t, yawQ);
+    }
+
+    doPointerDown(e) {
+        if (e.button === 2) this.service("InputManager").enterPointerLock();;
+    }
+
+    doPointerUp(e) {
+        if (e.button === 2) this.service("InputManager").exitPointerLock();
+    }
+
+    doPointerDelta(e) {
+        if (this.service("InputManager").inPointerLock) {
+            this.yaw += (-this.turnSpeed * e.xy[0]) % TAU;
+            this.pitch += (-this.turnSpeed * e.xy[1]) % TAU;
+            this.pitch = Math.max(-Math.PI/2, this.pitch);
+            this.pitch = Math.min(Math.PI/2, this.pitch);
+        };
+    }
+
+    get lookGlobal() {
+        const pitchQ = q_axisAngle([1,0,0], this.pitch);
+        const yawQ = q_axisAngle([0,1,0], this.yaw);
+        const lookQ = q_multiply(pitchQ, yawQ);
+
+        const local =  m4_scaleRotationTranslation(this.scale, lookQ, this.translation)
+        let global= local;
+        if (this.parent && this.parent.global) global = m4_multiply(local, this.parent.global);
+
+        return global;
     }
 
 }
+
+
 
 
 //------------------------------------------------------------------------------------------
@@ -263,9 +308,9 @@ class MyModelRoot extends ModelRoot {
         this.testActor = TestActor.create({translation: [0,0,-3]});
 
 
-        this.subscribe("input", "zDown", this.test0);
-        this.subscribe("input", "xDown", this.test1);
-        this.subscribe("input", "cDown", this.test2);
+        // this.subscribe("input", "zDown", this.test0);
+        // this.subscribe("input", "xDown", this.test1);
+        // this.subscribe("input", "cDown", this.test2);
     }
 
     test0() {
@@ -293,7 +338,7 @@ MyModelRoot.register("MyModelRoot");
 class MyViewRoot extends ViewRoot {
 
     static viewServices() {
-        return [InputManager, ThreeRenderManager, WidgetManager];
+        return [InputManager, UIManager, ThreeRenderManager, WidgetManager];
     }
 
     constructor(model) {
@@ -302,20 +347,52 @@ class MyViewRoot extends ViewRoot {
         three.renderer.setClearColor(new THREE.Color(0.45, 0.8, 0.8));
 
 
-        // const widget0 = new VisibleWidget3({color: [0,1,1]});
-        // const widget1 = new Widget3({parent: widget0});
-        // const widget2 = new Widget3({parent: widget0});
+        // const input = this.service("InputManager");
+        // const ui = this.service("UIManager");
+        // this.HUD = new Widget({parent: ui.root, autoSize: [1,1]});
+        // this.button0 = new ButtonWidget({
+        //     parent: this.HUD,
+        //     local: [20,20],
+        //     size: [100,50],
+        //     label: new TextWidget({point: 12, text: "MouseLook"}),
+        //     onClick: this.onClick
+        // });
 
-        // console.log(widget0.children);
-        // console.log(widget1);
-        // console.log(widget2);
+        // this.subscribe("input", "pointerDelta", this.onPointerDelta);
+        // this.subscribe("input", "zDown", this.startMouselook);
+        // this.subscribe("input", "zUp", this.stopMouselook);
 
-        // this.subscribe("input", "bDown", this.test);
-
-        // const xxx = {visible: true};
-        // console.log(xxx.visible);
-        // console.log(xxx.visible === undefined || xxx.visible);
     }
+
+    // onPointerDelta(e) {
+    //     const input = this.service("InputManager");
+    //     if (input.inPointerLock) {
+    //         console.log(e.xy);
+    //     };
+    //     // console.log(e.xy);
+    // }
+
+
+
+    // onClick(e) {
+    //     // console.log("click!");
+    //     const input = this.service("InputManager");
+    //     input.enterPointerLock();
+    // }
+
+    // startMouselook() {
+    //     console.log("start!");
+    //     const input = this.service("InputManager");
+    //     input.enterPointerLock();
+
+    // }
+
+    // stopMouselook() {
+    //     console.log("stop!");
+    //     const input = this.service("InputManager");
+    //     input.exitPointerLock();
+
+    // }
 
 
 }